requirements.txt
aiohttp==2.2.5
async-timeout==2.0.1
awscli==1.15.6
boto3==1.6.6
botocore==1.10.6
cffi==1.11.5
chardet==3.0.4
colorama==0.3.7
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.1
django-constrainedfilefield==3.1.3
django-heroku==0.3.1
django-storages==1.6.6
django-toolbelt==0.0.1
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
multidict==4.2.0
opuslib==2.0.0
pathlib==1.0.1
Pillow==5.0.0
psycopg2==2.7.4
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.7.2
python-magic==0.4.15
pytz==2018.4
PyYAML==3.12
redis==2.10.6
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
static3==0.7.0
virtualenv==15.1.0
websockets==3.4
whitenoise==3.3.1
yarl==1.1.1
